requirements.txt
numpy==1.26.0
pandas==2.1.1
python_dateutil==2.8.2
streamlit==1.27.2
